+++ packages/bundle/src/createCognitiveServicesSpeechServicesPonyfillFactory.ts
@@ -1,35 +1,11 @@
-import { AudioConfig } from 'microsoft-cognitiveservices-speech-sdk/distrib/lib/src/sdk/Audio/AudioConfig';
+import { AudioConfig } from 'microsoft-cognitiveservices-speech-sdk';
 import { WebSpeechPonyfillFactory } from 'botframework-webchat-api';
 import createPonyfill from 'web-speech-cognitive-services/lib/SpeechServices';
 
+import CognitiveServicesAudioOutputFormat from './types/CognitiveServicesAudioOutputFormat';
 import CognitiveServicesCredentials from './types/CognitiveServicesCredentials';
-
-type CognitiveServicesAudioOutputFormat =
-  | 'audio-16khz-128kbitrate-mono-mp3'
-  | 'audio-16khz-32kbitrate-mono-mp3'
-  | 'audio-16khz-64kbitrate-mono-mp3'
-  | 'audio-24khz-160kbitrate-mono-mp3'
-  | 'audio-24khz-48kbitrate-mono-mp3'
-  | 'audio-24khz-96kbitrate-mono-mp3'
-  | 'audio-48khz-192kbitrate-mono-mp3'
-  | 'audio-48khz-96kbitrate-mono-mp3'
-  | 'ogg-16khz-16bit-mono-opus'
-  | 'ogg-24khz-16bit-mono-opus'
-  | 'ogg-48khz-16bit-mono-opus'
-  | 'raw-16khz-16bit-mono-pcm'
-  | 'raw-16khz-16bit-mono-truesilk'
-  | 'raw-24khz-16bit-mono-pcm'
-  | 'raw-24khz-16bit-mono-truesilk'
-  | 'raw-48khz-16bit-mono-pcm'
-  | 'raw-8khz-8bit-mono-alaw'
-  | 'raw-8khz-8bit-mono-mulaw'
-  | 'riff-16khz-16bit-mono-pcm'
-  | 'riff-24khz-16bit-mono-pcm'
-  | 'riff-48khz-16bit-mono-pcm'
-  | 'riff-8khz-8bit-mono-alaw'
-  | 'riff-8khz-8bit-mono-mulaw'
-  | 'webm-16khz-16bit-mono-opus'
-  | 'webm-24khz-16bit-mono-opus';
+import CognitiveServicesTextNormalization from './types/CognitiveServicesTextNormalization';
+import createMicrophoneAudioConfigAndAudioContext from './speech/createMicrophoneAudioConfigAndAudioContext';
 
 export default function createCognitiveServicesSpeechServicesPonyfillFactory({
   audioConfig,
@@ -46,11 +22,11 @@ export default function createCognitiveServicesSpeechServicesPonyfillFactory({
   audioContext?: AudioContext;
   audioInputDeviceId?: string;
   credentials: CognitiveServicesCredentials;
-  enableTelemetry?: boolean;
+  enableTelemetry?: true;
   speechRecognitionEndpointId?: string;
   speechSynthesisDeploymentId?: string;
   speechSynthesisOutputFormat?: CognitiveServicesAudioOutputFormat;
-  textNormalization?: 'display' | 'itn' | 'lexical' | 'maskeditn';
+  textNormalization?: CognitiveServicesTextNormalization;
 }): WebSpeechPonyfillFactory {
   if (!window.navigator.mediaDevices && !audioConfig) {
     console.warn(
@@ -60,19 +36,22 @@ export default function createCognitiveServicesSpeechServicesPonyfillFactory({
     return () => ({});
   }
 
-  if (audioConfig && audioInputDeviceId) {
-    console.warn(
-      'botframework-webchat: "audioConfig" and "audioInputDeviceId" cannot be set at the same time; ignoring "audioInputDeviceId".'
-    );
-  }
+  if (audioConfig) {
+    audioInputDeviceId &&
+      console.warn(
+        'botframework-webchat: "audioConfig" and "audioInputDeviceId" cannot be set at the same time; ignoring "audioInputDeviceId".'
+      );
 
-  // WORKAROUND: We should prevent AudioContext object from being recreated because they may be blessed and UX-wise expensive to recreate.
-  //             In Cognitive Services SDK, if they detect the "end" function is falsy, they will not call "end" but "suspend" instead.
-  //             And on next recognition, they will re-use the AudioContext object.
-  if (!audioConfig) {
-    audioConfig = audioInputDeviceId
-      ? AudioConfig.fromMicrophoneInput(audioInputDeviceId)
-      : AudioConfig.fromDefaultMicrophoneInput();
+    audioContext &&
+      console.warn(
+        'botframework-webchat: "audioConfig" and "audioContext" cannot be set at the same time; ignoring "audioContext" for speech recognition.'
+      );
+  } else {
+    ({ audioConfig, audioContext } = createMicrophoneAudioConfigAndAudioContext({
+      audioContext,
+      audioInputDeviceId,
+      enableTelemetry
+    }));
   }
 
   return ({ referenceGrammarID } = {}) => {
@@ -89,6 +68,7 @@ export default function createCognitiveServicesSpeechServicesPonyfillFactory({
     });
 
     return {
+      resumeAudioContext: () => audioContext && audioContext.state === 'suspended' && audioContext.resume(),
       SpeechGrammarList,
       SpeechRecognition,
       speechSynthesis,