+++ packages/component/src/hooks/internal/useResumeAudioContext.ts
@@ -1,0 +1,7 @@
+import useWebSpeechPonyfill from '../useWebSpeechPonyfill';
+
+export default function useResumeAudioContext(): () => Promise<void> {
+  const [{ resumeAudioContext }] = useWebSpeechPonyfill();
+
+  return () => resumeAudioContext && resumeAudioContext();
+}